refactor: extract onboarding logic to custom hook
